requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.8.0
Werkzeug==2.3.7
win32-setctime==1.1.0
